refactor: add actionProcessor & scene4 demo
