+++ packages/vstory/demo/index.html
@@ -8,7 +8,8 @@
   </head>
 
   <body>
-    <div id="root"></div>
+    <!-- 视频 1920*1080, 等比例缩小 0.75 -->
+    <div id="root" style="width: 1440px; height: 810px"></div>
   </body>
   <script type="module" src="/src/App.tsx"></script>
 </html>
