+++ packages/vstory/src/dsl/story-processor/index.ts
@@ -1,1 +1,3 @@
+export * from './processor';
+
 export { processorMap } from './processorMap/processorMap';
